requirements.txt
Django==5.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
django-extensions==3.2.3
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
black==23.12.1
flake8==7.0.0
isort==5.13.2

